+++ .github/workflows/sandbox-runtime-tests.yml
@@ -17,7 +17,7 @@ jobs:
     defaults:
       run:
         shell: bash
-        working-directory: dify-agent-sandbox-runtime
+        working-directory: dify-agent-runtime
 
     steps:
       - name: Checkout code
@@ -29,8 +29,8 @@ jobs:
       - name: Setup Go
         uses: actions/setup-go@d35c59abb061a4a6fb18e82ac0862c26744d6ab5 # v5.5.0
         with:
-          go-version-file: dify-agent-sandbox-runtime/go.mod
-          cache-dependency-path: dify-agent-sandbox-runtime/go.sum
+          go-version-file: dify-agent-runtime/go.mod
+          cache-dependency-path: dify-agent-runtime/go.sum
 
       - name: Run unit tests
         run: go test -race -count=1 ./...
@@ -41,7 +41,7 @@ jobs:
     defaults:
       run:
         shell: bash
-        working-directory: dify-agent-sandbox-runtime
+        working-directory: dify-agent-runtime
 
     steps:
       - name: Checkout code
@@ -53,13 +53,13 @@ jobs:
       - name: Setup Go
         uses: actions/setup-go@d35c59abb061a4a6fb18e82ac0862c26744d6ab5 # v5.5.0
         with:
-          go-version-file: dify-agent-sandbox-runtime/go.mod
-          cache-dependency-path: dify-agent-sandbox-runtime/go.sum
+          go-version-file: dify-agent-runtime/go.mod
+          cache-dependency-path: dify-agent-runtime/go.sum
 
       - name: Run golangci-lint
         uses: golangci/golangci-lint-action@4afd733a84b1f43292c63897423277bb7f4313a9 # v6.5.0
         with:
-          working-directory: dify-agent-sandbox-runtime
+          working-directory: dify-agent-runtime
           version: latest
 
   sandbox-runtime-integration:
@@ -68,7 +68,7 @@ jobs:
     defaults:
       run:
         shell: bash
-        working-directory: dify-agent-sandbox-runtime
+        working-directory: dify-agent-runtime
 
     steps:
       - name: Checkout code
@@ -80,8 +80,8 @@ jobs:
       - name: Setup Go
         uses: actions/setup-go@d35c59abb061a4a6fb18e82ac0862c26744d6ab5 # v5.5.0
         with:
-          go-version-file: dify-agent-sandbox-runtime/go.mod
-          cache-dependency-path: dify-agent-sandbox-runtime/go.sum
+          go-version-file: dify-agent-runtime/go.mod
+          cache-dependency-path: dify-agent-runtime/go.sum
 
       - name: Build and start runtime
         run: make integration-up
