+++ .gitignore
@@ -263,6 +263,6 @@ scripts/stress-test/reports/
 # Vitest local reports
 web/.vitest-reports/
 
-# dify-agent-sandbox-runtime
-dify-agent-sandbox-runtime/bin/
-dify-agent-sandbox-runtime/.integration-state
+# dify-agent-runtime
+dify-agent-runtime/bin/
+dify-agent-runtime/.integration-state
