+++ dify-agent-runtime/cmd/dify-agent-cli/main.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/langgenius/dify/dify-agent-sandbox-runtime/internal/agentcli"
+	"github.com/langgenius/dify/dify-agent-runtime/internal/agentcli"
 )
 
 func main() {
