+++ dify-agent-runtime/cmd/shellctl/main.go
@@ -13,7 +13,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/langgenius/dify/dify-agent-sandbox-runtime/internal/server"
+	"github.com/langgenius/dify/dify-agent-runtime/internal/server"
 )
 
 func main() {
